prov/opx: Move dput packet build functions to new inline include.
Move all the inline write_header_and_payload functions in fi_opx_hfi1.c
into fi_opx_hfi1_inlines.h (new include).

Move fi_opx_hfi1_dput_sdma_pending_completion out of fi_opx_hfi1.c and into
fi_opx_hfi1_sdma.c.

This commit does not add any new function nor change any existing function.
