+++ prov/opx/Makefile.include
@@ -120,6 +120,7 @@ noinst_HEADERS = \
 	$(top_srcdir)/prov/opx/include/rdma/opx/fi_opx_hfi1_packet.h \
 	$(top_srcdir)/prov/opx/include/rdma/opx/fi_opx_hfi1.h \
 	$(top_srcdir)/prov/opx/include/rdma/opx/fi_opx_hfi1_fabric.h \
+	$(top_srcdir)/prov/opx/include/rdma/opx/fi_opx_hfi1_inlines.h \
 	$(top_srcdir)/prov/opx/include/rdma/opx/fi_opx_internal.h \
 	$(top_srcdir)/prov/opx/include/rdma/opx/fi_opx_rma_ops.h \
 	$(top_srcdir)/prov/opx/include/rdma/opx/fi_opx_compiler.h \
