+++ prov/opx/src/fi_opx_hfi1.c
@@ -37,6 +37,7 @@
 
 #include "rdma/fabric.h" // only for 'fi_addr_t' ... which is a typedef to uint64_t
 #include "rdma/opx/fi_opx_hfi1.h"
+#include "rdma/opx/fi_opx_hfi1_inlines.h"
 #include "rdma/opx/fi_opx.h"
 #include "rdma/opx/fi_opx_eq.h"
 #include "rdma/opx/fi_opx_hfi1_sdma.h"
@@ -1400,325 +1401,6 @@ void opx_hfi1_dput_fence(struct fi_opx_ep *opx_ep,
 	slist_insert_tail(&work->work_elem.slist_entry, &opx_ep->tx->work_pending);
 }
 
-__OPX_FORCE_INLINE__
-size_t opx_hfi1_dput_write_header_and_payload_put(
-				struct fi_opx_ep *opx_ep,
-				union fi_opx_hfi1_packet_hdr *tx_hdr,
-				union fi_opx_hfi1_packet_payload *tx_payload,
-				struct iovec *iov,
-				const uint64_t op64,
-				const uint64_t dt64,
-				const size_t payload_bytes,
-				const uint64_t key,
-				uint8_t **sbuf,
-				uintptr_t *rbuf)
-{
-	tx_hdr->qw[4] = opx_ep->rx->tx.dput.hdr.qw[4] | FI_OPX_HFI_DPUT_OPCODE_PUT |
-			(dt64 << 16) | (op64 << 24) | (payload_bytes << 48);
-	tx_hdr->qw[5] = key;
-	tx_hdr->qw[6] = fi_opx_dput_rbuf_out(*rbuf);
-
-	if (tx_payload) {
-		assert(!iov);
-		memcpy((void *)tx_payload, (const void *)*sbuf, payload_bytes);
-	} else {
-		assert(iov);
-		iov->iov_base = (void *) *sbuf;
-		iov->iov_len = payload_bytes;
-	}
-
-	(*sbuf) += payload_bytes;
-	(*rbuf) += payload_bytes;
-
-	return payload_bytes;
-}
-
-__OPX_FORCE_INLINE__
-void opx_hfi1_dput_write_payload_atomic_fetch(
-				union fi_opx_hfi1_packet_payload *tx_payload,
-				const uint64_t dput_bytes,
-				const uint64_t fetch_vaddr,
-				const uintptr_t target_byte_counter_vaddr,
-				uint64_t bytes_sent,
-				uint8_t *sbuf)
-{
-	struct fi_opx_hfi1_dput_fetch *dput_fetch = (struct fi_opx_hfi1_dput_fetch *) tx_payload->byte;
-	dput_fetch->fetch_rbuf = fetch_vaddr + bytes_sent;
-	dput_fetch->fetch_counter_vaddr = target_byte_counter_vaddr;
-
-	memcpy((void *)&tx_payload->byte[sizeof(*dput_fetch)], (const void *)sbuf, dput_bytes);
-}
-
-__OPX_FORCE_INLINE__
-size_t opx_hfi1_dput_write_header_and_payload_atomic_fetch(
-				struct fi_opx_ep *opx_ep,
-				union fi_opx_hfi1_packet_hdr *tx_hdr,
-				union fi_opx_hfi1_packet_payload *tx_payload,
-				struct iovec *iov,
-				const uint64_t op64,
-				const uint64_t dt64,
-				const size_t payload_bytes,
-				const uint64_t key,
-				const uint64_t fetch_vaddr,
-				const uintptr_t target_byte_counter_vaddr,
-				uint64_t bytes_sent,
-				uint8_t **sbuf,
-				uintptr_t *rbuf)
-{
-	tx_hdr->qw[4] = opx_ep->rx->tx.dput.hdr.qw[4] | FI_OPX_HFI_DPUT_OPCODE_ATOMIC_FETCH |
-			(dt64 << 16) | (op64 << 24) | (payload_bytes << 48);
-	tx_hdr->qw[5] = key;
-	tx_hdr->qw[6] = fi_opx_dput_rbuf_out(*rbuf);
-
-	size_t dput_bytes = payload_bytes - sizeof(struct fi_opx_hfi1_dput_fetch);
-
-	if (tx_payload) {
-		assert(!iov);
-		opx_hfi1_dput_write_payload_atomic_fetch(tx_payload, dput_bytes,
-					fetch_vaddr, target_byte_counter_vaddr,
-					bytes_sent, *sbuf);
-
-		/* Here the source buffer is the data with no fi_opx_hfi1_dput_fetch
-		   contained in it, so we only want to advance the pointer by
-		   the number of actual data bytes, not including the dput fetch
-		   header info that we injected at the beginning in the target buffer. */
-		(*sbuf) += dput_bytes;
-	} else {
-		assert(iov);
-		iov->iov_base = (void *) *sbuf;
-		iov->iov_len = payload_bytes;
-
-		/* Here we're simply setting the IOV pointer, so the source buffer
-		   must already be a fully built packet payload, so advance the
-		   pointer by the full payload amount */
-		(*sbuf) += payload_bytes;
-	}
-
-	(*rbuf) += dput_bytes;
-
-	return dput_bytes;
-}
-
-__OPX_FORCE_INLINE__
-void opx_hfi1_dput_write_payload_atomic_compare_fetch(
-				union fi_opx_hfi1_packet_payload *tx_payload,
-				const size_t dput_bytes_half,
-				const uint64_t fetch_vaddr,
-				const uintptr_t target_byte_counter_vaddr,
-				uint64_t bytes_sent,
-				uint8_t *sbuf,
-				uint8_t *cbuf)
-{
-	struct fi_opx_hfi1_dput_fetch *dput_fetch = (struct fi_opx_hfi1_dput_fetch *) tx_payload->byte;
-	dput_fetch->fetch_rbuf = fetch_vaddr + bytes_sent;
-	dput_fetch->fetch_counter_vaddr = target_byte_counter_vaddr;
-
-	/* The first 1/2 of the actual payload bytes contains the data for the elements
-	   we want to write to memory at the destination. The second 1/2 contains the
-	   data for the elements to use as compare values against the elements currently
-	   in the destination's memory, to see if a swap should take place. */
-	memcpy((void *)&tx_payload->byte[sizeof(*dput_fetch)],
-		(const void *)sbuf, dput_bytes_half);
-	memcpy((void *)&tx_payload->byte[sizeof(*dput_fetch) + dput_bytes_half],
-		(const void *)cbuf, dput_bytes_half);
-}
-
-__OPX_FORCE_INLINE__
-size_t opx_hfi1_dput_write_header_and_payload_atomic_compare_fetch(
-				struct fi_opx_ep *opx_ep,
-				union fi_opx_hfi1_packet_hdr *tx_hdr,
-				union fi_opx_hfi1_packet_payload *tx_payload,
-				struct iovec *iov,
-				const uint64_t op64,
-				const uint64_t dt64,
-				const size_t payload_bytes,
-				const uint64_t key,
-				const uint64_t fetch_vaddr,
-				const uintptr_t target_byte_counter_vaddr,
-				uint64_t bytes_sent,
-				uint8_t **sbuf,
-				uint8_t **cbuf,
-				uintptr_t *rbuf)
-{
-	tx_hdr->qw[4] = opx_ep->rx->tx.dput.hdr.qw[4] | FI_OPX_HFI_DPUT_OPCODE_ATOMIC_COMPARE_FETCH |
-			(dt64 << 16) | (op64 << 24) | (payload_bytes << 48);
-	tx_hdr->qw[5] = key;
-	tx_hdr->qw[6] = fi_opx_dput_rbuf_out(*rbuf);
-
-	size_t dput_bytes = payload_bytes - sizeof(struct fi_opx_hfi1_dput_fetch);
-	size_t dput_bytes_half = dput_bytes >> 1;
-
-	if (tx_payload) {
-		assert(!iov);
-		opx_hfi1_dput_write_payload_atomic_compare_fetch(tx_payload, dput_bytes_half,
-					fetch_vaddr, target_byte_counter_vaddr,
-					bytes_sent, *sbuf, *cbuf);
-
-		/* Here the source buffer is the data with no fi_opx_hfi1_dput_fetch
-		   contained in it, so we only want to advance the pointer by
-		   the number of actual data bytes, not including the dput fetch
-		   header info that we injected at the beginning in the target buffer. */
-		(*sbuf) += dput_bytes_half;
-	} else {
-		assert(iov);
-		iov->iov_base = (void *) *sbuf;
-		iov->iov_len = payload_bytes;
-
-		/* Here we're simply setting the IOV pointer, so the source buffer
-		   must already be a fully built packet payload, so advance the
-		   pointer by the full payload amount */
-		(*sbuf) += payload_bytes;
-	}
-
-	(*rbuf) += dput_bytes_half;
-	(*cbuf) += dput_bytes_half;
-
-	return dput_bytes;
-}
-
-__OPX_FORCE_INLINE__
-size_t opx_hfi1_dput_write_header_and_payload_get(
-				struct fi_opx_ep *opx_ep,
-				union fi_opx_hfi1_packet_hdr *tx_hdr,
-				union fi_opx_hfi1_packet_payload *tx_payload,
-				struct iovec *iov,
-				const uint64_t dt64,
-				const size_t payload_bytes,
-				const uintptr_t target_byte_counter_vaddr,
-				uint8_t **sbuf,
-				uintptr_t *rbuf)
-{
-	tx_hdr->qw[4] = opx_ep->rx->tx.dput.hdr.qw[4] | FI_OPX_HFI_DPUT_OPCODE_GET |
-			(dt64 << 16) | (payload_bytes << 48);
-	tx_hdr->qw[5] = target_byte_counter_vaddr;
-	tx_hdr->qw[6] = fi_opx_dput_rbuf_out(*rbuf);
-
-	if (tx_payload) {
-		assert(!iov);
-		if (dt64 == (FI_VOID - 1)) {
-			memcpy((void *)tx_payload, (const void *)*sbuf, payload_bytes);
-		} else {
-			fi_opx_rx_atomic_dispatch((void *)*sbuf,
-						  (void *)tx_payload,
-						  payload_bytes,
-						  dt64,
-						  FI_ATOMIC_WRITE);
-		}
-	} else {
-		assert(iov);
-		iov->iov_base = (void *) *sbuf;
-		iov->iov_len = payload_bytes;
-	}
-
-	(*sbuf) += payload_bytes;
-	(*rbuf) += payload_bytes;
-
-	return payload_bytes;
-}
-
-__OPX_FORCE_INLINE__
-size_t opx_hfi1_dput_write_header_and_payload_rzv(
-				struct fi_opx_ep *opx_ep,
-				union fi_opx_hfi1_packet_hdr *tx_hdr,
-				union fi_opx_hfi1_packet_payload *tx_payload,
-				struct iovec *iov,
-				const uint64_t op64,
-				const uint64_t dt64,
-				const size_t payload_bytes,
-				const uint32_t opcode,
-				const uintptr_t target_byte_counter_vaddr,
-				uint8_t **sbuf,
-				uintptr_t *rbuf)
-{
-	tx_hdr->qw[4] = opx_ep->rx->tx.dput.hdr.qw[4] | (opcode) | (payload_bytes << 48);
-	tx_hdr->qw[5] = target_byte_counter_vaddr;
-	tx_hdr->qw[6] = fi_opx_dput_rbuf_out(*rbuf);
-
-	if (tx_payload) {
-		assert(!iov);
-		memcpy((void *)tx_payload, (const void *)*sbuf, payload_bytes);
-	} else {
-		assert(iov);
-		iov->iov_base = (void *) *sbuf;
-		iov->iov_len = payload_bytes;
-	}
-	(*sbuf) += payload_bytes;
-	(*rbuf) += payload_bytes;
-
-	return payload_bytes;
-}
-
-__OPX_FORCE_INLINE__
-size_t opx_hfi1_dput_write_header_and_payload(
-				struct fi_opx_ep *opx_ep,
-				union fi_opx_hfi1_packet_hdr *tx_hdr,
-				union fi_opx_hfi1_packet_payload *tx_payload,
-				struct iovec *iov,
-				const uint32_t opcode,
-				const int64_t psn_orig,
-				const uint16_t lrh_dws,
-				const uint64_t op64,
-				const uint64_t dt64,
-				const uint64_t lrh_dlid,
-				const uint64_t bth_rx,
-				const size_t payload_bytes,
-				const uint64_t key,
-				const uint64_t fetch_vaddr,
-				const uintptr_t target_byte_counter_vaddr,
-				uint64_t bytes_sent,
-				uint8_t **sbuf,
-				uint8_t **cbuf,
-				uintptr_t *rbuf)
-{
-	uint64_t psn = (uint64_t) htonl((uint32_t)psn_orig);
-
-	tx_hdr->qw[0] = opx_ep->rx->tx.dput.hdr.qw[0] | lrh_dlid | ((uint64_t)lrh_dws << 32);
-	tx_hdr->qw[1] = opx_ep->rx->tx.dput.hdr.qw[1] | bth_rx;
-	tx_hdr->qw[2] = opx_ep->rx->tx.dput.hdr.qw[2] | psn;
-	tx_hdr->qw[3] = opx_ep->rx->tx.dput.hdr.qw[3];
-
-	switch(opcode) {
-	case FI_OPX_HFI_DPUT_OPCODE_RZV:
-	case FI_OPX_HFI_DPUT_OPCODE_RZV_TID:
-	case FI_OPX_HFI_DPUT_OPCODE_RZV_NONCONTIG:
-		return opx_hfi1_dput_write_header_and_payload_rzv(
-				opx_ep, tx_hdr, tx_payload, iov,
-				op64, dt64, payload_bytes, opcode,
-				target_byte_counter_vaddr, sbuf, rbuf);
-		break;
-	case FI_OPX_HFI_DPUT_OPCODE_GET:
-		return opx_hfi1_dput_write_header_and_payload_get(
-				opx_ep, tx_hdr, tx_payload, iov,
-				dt64, payload_bytes,
-				target_byte_counter_vaddr, sbuf, rbuf);
-		break;
-	case FI_OPX_HFI_DPUT_OPCODE_PUT:
-		return opx_hfi1_dput_write_header_and_payload_put(
-				opx_ep, tx_hdr, tx_payload,
-				iov, op64, dt64, payload_bytes,
-				key, sbuf, rbuf);
-		break;
-	case FI_OPX_HFI_DPUT_OPCODE_ATOMIC_FETCH:
-		return opx_hfi1_dput_write_header_and_payload_atomic_fetch(
-				opx_ep, tx_hdr, tx_payload, iov, op64, dt64,
-				payload_bytes, key, fetch_vaddr,
-				target_byte_counter_vaddr,
-				bytes_sent, sbuf, rbuf);
-		break;
-	case FI_OPX_HFI_DPUT_OPCODE_ATOMIC_COMPARE_FETCH:
-		return opx_hfi1_dput_write_header_and_payload_atomic_compare_fetch(
-				opx_ep, tx_hdr, tx_payload, iov, op64, dt64,
-				payload_bytes, key, fetch_vaddr,
-				target_byte_counter_vaddr,
-				bytes_sent, sbuf, cbuf, rbuf);
-		break;
-	default:
-		FI_WARN(fi_opx_global.prov, FI_LOG_EP_DATA,
-			"Invalid opcode %0X; abort\n", opcode);
-		abort();
-	}
-}
-
 int fi_opx_hfi1_do_dput (union fi_opx_hfi1_deferred_work * work)
 {
 	struct fi_opx_hfi1_dput_params *params = &work->dput;
@@ -1928,69 +1610,6 @@ int fi_opx_hfi1_do_dput (union fi_opx_hfi1_deferred_work * work)
 	} /* for niov */
 
 	params->work_elem.complete = true;
-	return FI_SUCCESS;
-}
-
-int fi_opx_hfi1_dput_sdma_pending_completion(union fi_opx_hfi1_deferred_work *work)
-{
-	struct fi_opx_hfi1_dput_params *params = &work->dput;
-	struct fi_opx_ep * opx_ep = params->opx_ep;
-
-	assert(params->work_elem.low_priority);
-
-	fi_opx_hfi1_sdma_poll_completion(opx_ep);
-
-	struct fi_opx_hfi1_sdma_work_entry *we = (struct fi_opx_hfi1_sdma_work_entry *) params->sdma_reqs.head;
-	while (we) {
-		// If we're using the SDMA WE bounce buffer, we need to wait for
-		// the hit_zero to mark the work element as complete. The replay
-		// iovecs are pointing to the SDMA WE bounce buffers, so we can't
-		// free the SDMA WEs until the replays are cleared.
-		if (!params->work_elem.complete && we->use_bounce_buf) {
-			FI_OPX_DEBUG_COUNTERS_INC(work->dput.opx_ep->debug_counters.sdma.eagain_pending_dc);
-			return -FI_EAGAIN;
-		}
-		enum hfi1_sdma_comp_state we_status = fi_opx_hfi1_sdma_get_status(opx_ep, we);
-		if (we_status == QUEUED) {
-			FI_OPX_DEBUG_COUNTERS_INC(opx_ep->debug_counters.sdma.eagain_pending_writev);
-			FI_DBG_TRACE(fi_opx_global.prov, FI_LOG_EP_DATA, "FI_EAGAIN\n");
-			return -FI_EAGAIN;
-		}
-		assert(we_status == COMPLETE);
-
-		slist_remove_head(&params->sdma_reqs);
-		we->next = NULL;
-		fi_opx_hfi1_sdma_return_we(params->opx_ep, we);
-		we = (struct fi_opx_hfi1_sdma_work_entry *) params->sdma_reqs.head;
-	}
-
-	assert(slist_empty(&params->sdma_reqs));
-
-	if (!params->work_elem.complete) {
-		assert(params->delivery_completion);
-		FI_OPX_DEBUG_COUNTERS_INC(work->dput.opx_ep->debug_counters.sdma.eagain_pending_dc);
-		return -FI_EAGAIN;
-	}
-
-	if (params->origin_byte_counter) {
-		// If we're not doing delivery_competion, then origin_byte_counter
-		// should have already been zero'd and NULL'd at the end of do_dput_sdma(...)
-		assert(params->delivery_completion);
-		*params->origin_byte_counter = 0;
-		params->origin_byte_counter = NULL;
-	}
-
-	if (params->user_cc) {
-		assert(params->user_cc->byte_counter >= params->cc->initial_byte_count);
-		params->user_cc->byte_counter -= params->cc->initial_byte_count;
-		if (params->user_cc->byte_counter == 0) {
-			params->user_cc->hit_zero(params->user_cc);
-		}
-	}
-	OPX_BUF_FREE(params->cc);
-
-	FI_DBG_TRACE(fi_opx_global.prov, FI_LOG_EP_DATA,
-		"===================================== PENDING DPUT %u COMPLETE\n", work->work_elem.complete);
 	return FI_SUCCESS;
 }
